Enable rubocop and fix most of the issues

+++ source/linkmunger.rb
@@ -1,4 +1,6 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
+
 # USAGE:
 #  ./linkmunger.rb [-n] <PATH>
 # Takes a path to the folder acting as site root as argument, or else
@@ -8,64 +10,61 @@
 
 def relativepath(path, relative_to)
   if path == relative_to
-    if path[-1] == "/" and relative_to[-1] == "/"
-      return "."
-    else
-      return File.basename(path)
-    end
+    return '.' if (path[-1] == '/') && (relative_to[-1] == '/')
+
+    return File.basename(path)
+
   end
 
   if path == "#{relative_to}/"
     # from /a to /a/
-    return File.basename(path) + "/"
+    return File.basename(path) + '/'
   end
 
   if "#{path}/" == relative_to
     # from /a/ to /a
-    return "../" + File.basename(path)
+    return "../#{File.basename(path)}"
   end
 
-  path = path.chomp("/")
-  if path == ""
-    path = [""]
-  else
-    # The negative limit ensures that empty values aren't dropped.
-    path = path.split(File::SEPARATOR, -1)
-  end
+  path = path.chomp('/')
+  path = if path == ''
+           ['']
+         else
+           # The negative limit ensures that empty values aren't dropped.
+           path.split(File::SEPARATOR, -1)
+         end
 
-  relative_to = relative_to.chomp("/")
-  if relative_to == ""
-    relative_to = [""]
-  else
-    # The negative limit ensures that empty values aren't dropped.
-    relative_to = relative_to.split(File::SEPARATOR, -1)
-  end
+  relative_to = relative_to.chomp('/')
+  relative_to = if relative_to == ''
+                  ['']
+                else
+                  # The negative limit ensures that empty values aren't dropped.
+                  relative_to.split(File::SEPARATOR, -1)
+                end
 
-  while (path.length > 0) && (path.first == relative_to.first)
+  while path.length.positive? && (path.first == relative_to.first)
     path.shift
     relative_to.shift
   end
 
-  if relative_to.length == 0 and path.length == 0
-    throw "BUG: Processed paths were equivalent when raw paths were the same"
-  elsif relative_to.length == 0
+  if relative_to.empty? && path.empty?
+    throw 'BUG: Processed paths were equivalent when raw paths were the same'
+  elsif relative_to.empty?
     path.join(File::SEPARATOR)
-  elsif path.length == 0 and relative_to.length == 1
-    "."
+  elsif path.empty? && (relative_to.length == 1)
+    '.'
   else
-    (['..'] * (relative_to.length - 1) + path).join(File::SEPARATOR)
+    ((['..'] * (relative_to.length - 1)) + path).join(File::SEPARATOR)
   end
 end
 
-noop = ARGV.first == "-n"
+noop = ARGV.first == '-n'
 if noop
   ARGV.shift
-  puts "Running in no-op mode"
+  puts 'Running in no-op mode'
 end
 
-if ARGV.first
-  Dir.chdir(ARGV.shift)
-end
+Dir.chdir(ARGV.shift) if ARGV.first
 
 regex = %r{
   (href|src)=
@@ -88,14 +87,13 @@ Dir['**/*.html'].each do |page_path|
     prev_dir = File.dirname(page_path)
     puts "Processing: #{prev_dir}/"
   end
-  results = File.read(page_path).gsub(regex) do |match|
+  results = File.read(page_path).gsub(regex) do |_match|
     changed += 1
-    "#{$1}=#{$2}" + relativepath($3, "/" + page_path) + $2
+    "#{Regexp.last_match(1)}=#{Regexp.last_match(2)}" + relativepath(Regexp.last_match(3),
+                                                                     "/#{page_path}") + Regexp.last_match(2)
   end
 
-  if ! noop
-    File.write(page_path, results)
-  end
+  File.write(page_path, results) unless noop
 end
 
 puts "Updated #{changed} links"
